requirements.txt
idna
discord.py==2.3.2
dnspython==1.16.0
PyNaCl
async-timeout==3.0.1
pandas
gspread
oauth2client
psycopg2-binary
pydrive2
dpymenus
google-api-python-client
asyncpraw==7.7.1
thefuzz